types: introduce Fault as an attribute in a condition
This is for development and testing purposes, and enables the caller to insert
a fault into the controller when its executing on a condition.

Its upto the controller to implement and act on the fault, the callers
must take precautions to ensure faults are not acted on while running in
production.