remove the single-condition create method
We updated the CreateMultiple method to take a facility parameter to
assist in reconciliation. However, Create was not updated, which leads
to a ConditionRecord with no facility set. Seeing as CreateMultiple is
variadic, there's no real reason to keep a single condition version
around (although arguably CreatMultiple should be renamed).