event handler refactor (#232)
* WIP event handler refactor: needs tests

* orchestrator/reconcile: Fixes incorrect reconcile of Condition in queue

The orchestrator would incorrectly identify a Condition for reconcile
while the Condition was still in queue.

The Condition.UpdatedAt value is now checked and only if it exceeds the
queue max age threshold will it be marked for reconcile.

Updates test cases

* store/nats: ConditionRecord struct to include server facility

This is to enable the orchestrator to differentiate between Condition
Records that it is responsible for

* store/mock: gen-mock for updated interface

* remove the single-condition create method

We updated the CreateMultiple method to take a facility parameter to
assist in reconciliation. However, Create was not updated, which leads
to a ConditionRecord with no facility set. Seeing as CreateMultiple is
variadic, there's no real reason to keep a single condition version
around (although arguably CreatMultiple should be renamed).

* restrict incomplete records to this orchestrator's facility

* adjust client test expectations

* ensure update kind matches condition kind in merge

* add some behavioral tests for queuing next conditions

* clean up deprecated config warnings

* bump linter version

* delete the status record if it can not be reconciled

---------

Co-authored-by: Joel Rebello <jrebello@packet.com>