api/v1/orc: update handler conditionPending() -> conditionGet()
Updates the handler to return an incomplete condition which could be in
Pending or Active states. This enables the client to make decisions on
what steps it should take next.

The previous implementation would only return a Pending condition, this
would limit the client from taking decisions where it would have had to resume
an active condition.