add the firmware validation condition
Previous work added an endpoint in Condition API to queue work for
firmware validation but did not add an actual condition to trigger a
controller to perform the validation step, because the controller
support was not implemented. Now that this support exists, it is time to
integrate the actual firmware validation task.