api/conditions/handlers: Condition to include TraceID, SpanID, Client ID attributes
This enables tracing and linking Conditions executed by Controllers and
managed by the Orchestrator. So far, Span, TraceIDs were included in the
the NATS message header when a Condition gets published to the
Jetstream. Since we have Conditions that not published onto the
Jetstream and we'd like to ensure the traces are linked, the TraceID,
SpanID is incuded in the Condition payload itself and the NATS message
header can be deprecated.

The Client ID information is included to ensure we can identify the
client that queued the request.