keep status value after completion (#275)
Deleting the status value that a controller uses to track its work can
lead to that controller continuing to work after the orchestrator's
reconciler goroutine has marked the condition as failed. In a situation
where the controller attempts a status update and the orchestrator has
already failed the condition, the only acceptable response is for the
controller to stop work.